Created shared project for repository related stuff, moved the common stuff into said repository
