--- ExpressedRealms.Repositories.Characters/ExternalDependencies/IUserContext.cs
@@ -1,6 +1,0 @@
-namespace ExpressedRealms.Repositories.Characters.ExternalDependencies;
-
-public interface IUserContext
-{
-    public string CurrentUserId();
-}
