+++ ExpressedRealms.Repositories.Shared/ExternalDependencies/IUserContext.cs
@@ -1,0 +1,6 @@
+namespace ExpressedRealms.Repositories.Shared.ExternalDependencies;
+
+public interface IUserContext
+{
+    public string CurrentUserId();
+}
