+++ ExpressedRealms.sln
@@ -13,6 +13,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExpressedRealms.Email", "Ex
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExpressedRealms.Repositories.Characters", "ExpressedRealms.Repositories.Characters\ExpressedRealms.Repositories.Characters.csproj", "{427CB02E-A9FA-4308-B96D-B8BE2F1EB316}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExpressedRealms.Repositories.Shared", "ExpressedRealms.Repositories.Shared\ExpressedRealms.Repositories.Shared.csproj", "{0554641F-EEE1-4480-925F-F6FAB1993D99}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -41,6 +43,10 @@ Global
 		{427CB02E-A9FA-4308-B96D-B8BE2F1EB316}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{427CB02E-A9FA-4308-B96D-B8BE2F1EB316}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{427CB02E-A9FA-4308-B96D-B8BE2F1EB316}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0554641F-EEE1-4480-925F-F6FAB1993D99}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0554641F-EEE1-4480-925F-F6FAB1993D99}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0554641F-EEE1-4480-925F-F6FAB1993D99}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0554641F-EEE1-4480-925F-F6FAB1993D99}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
